fix: vue not rendering detailed post at /post/:id
I was encountering an issue with Vue when rendering the detailed post.
It was only showing the h1 tag. When opening up the console with the Vue
devtools extension, I encountered this error:

`Component template should contain exactly one root element`.

As it turns out, you only only have one root element in a template. What
that essentially meant in this circumstance is that I had to wrap
everything in <div> tags. I thus moved the <div id="post-long"> into the
<detailedpost> template, thereby solving the issue.