fix: add comment privacy if users don't follow author
If users visited a private post made by an author whose profile was
private, they were still able to see the comments because of the way the
`getPostComments` function worked.

This commit fixes that by only sending the comments if they follow the
author (if the author's settings are private). Otherwise, it will send
an error.