fix: correctly display comments
It's better to assign the comments their own array within `postDetails`.
Fixed the vue display in post.html.

TODO: error in server/queries.js in queryPostComments and
getCommentDetailed. It is prematurely returning the comments, so some
are missing when the response is sent.

TODO: in the frontend, the names of the authors of comments are not
displaying.